+++ rules/errors.js
@@ -48,6 +48,8 @@ module.exports = {
 		'no-sparse-arrays': 2,
 		// disallow unreachable statements after a return, throw, continue, or break statement
 		'no-unreachable': 2,
+		// disallow control flow statements in finally blocks
+		'no-unsafe-finally': 2,
 		// disallow comparisons with the value NaN
 		'use-isnan': 2,
 		// ensure JSDoc comments are valid