+++ rules/variables.js
@@ -14,10 +14,10 @@ module.exports = {
 		'no-shadow-restricted-names': 2,
 		// disallow declaration of variables already declared in the outer scope
 		'no-shadow': 2,
-		// disallow use of undefined when initializing variables
-		'no-undef-init': 0,
 		// disallow use of undeclared variables unless mentioned in a /*global */ block
 		'no-undef': 2,
+		// disallow use of undefined when initializing variables
+		'no-undef-init': 0,
 		// disallow use of undefined variable
 		'no-undefined': 0,
 		// disallow declaration of variables that are not used in the code