+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 2.6.0
+- Added errors/no-unsafe-finally rule
+- Added es6/no-duplicate-imports rule
+- Added es6/no-useless-computed-key rule
+- Added es6/no-useless-constructor rule
+- Added es6/prefer-arrow-callback rule
+- Added es6/prefer-template rule
+- Added es6/template-curly-spacing rule
+- Added es6/yield-star-spacing rule
+
 ## 2.5.0
 - Updated to use eslint-plugin-react 5.x